Move link component to ui folder
