+++ site/gatsby-site/src/pages/apps/classifications.js
@@ -8,7 +8,7 @@ import config from '../../../config';
 
 import { useTable, useFilters, usePagination, useSortBy } from 'react-table';
 import { Table, Spinner, Form, InputGroup, FormControl, Button } from 'react-bootstrap';
-import Link from 'components/Link';
+import Link from 'components/ui/Link';
 import { faExpandAlt } from '@fortawesome/free-solid-svg-icons';
 import { faEdit } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
