+++ site/gatsby-site/src/pages/summaries/incidents.js
@@ -5,7 +5,7 @@ import { graphql } from 'gatsby';
 import Button from 'react-bootstrap/Button';
 
 import Layout from 'components/Layout';
-import Link from 'components/Link';
+import Link from 'components/ui/Link';
 import { StyledHeading, StyledMainWrapper } from 'components/styles/Docs';
 
 const ReportList = ({ items }) => {
